API

Create new grids

AdaptiveDensityApproximation.create_grid — Function

create_grid(axis_ticks::AbstractArray...;initial_weight = 1.0, exclude_strings = [""],string_length = 10)

Create a multidimensional grid, where the axis_ticks define the corner points of the blocks.

For each dimension a separate array of axis-ticks is required. Each axis-ticks array must have at least 3 elements.

Keywords

initial_weight: Initial weight for the blocks.

exclude_strings: Strings that should not be used for the block names / keys.

string_length: Length for the block names / keys.

create_grid(axis_ticks::AbstractArray;initial_weight = 1.0, exclude_strings = [""],string_length = 10)

Create a one-dimensional grid where the axis_ticks define the start/end points of the intervals.

There must be at least 3 elements in the axis_ticks array.

Keywords

initial_weight: Initial weight for the intervals.

exclude_strings: Strings that should not be used for the block names / keys.

string_length: Length for the block names / keys.

Approximate density functions

AdaptiveDensityApproximation.approximate_density! — Function

approximate_density!(grid,f::Function; mode = :center, mesh_size = 4, volume_normalization = false)

Approximate the density function f with the grid, changing the weights to the function values and return the mutated grid.

mode = :center: Evaluate the density function in the center of the interval/block.

mode = :mean: Evaluate the density function at all corner points of the interval/block and use the mean value.

mode = :mesh: Evaluate the density function at all mesh points of the interval/block and use the mean value.

mesh_size = 4: Number of block discretization points in each dimension. Only applicable to mode = :mesh.

volume_normalization = false: If true the density value is normalized to the block volume (weight = value × volume).

Import and Export

AdaptiveDensityApproximation.export_weights — Function

export_weights(grid)

Return a one-dim array containing the weights of the intervals/blocks.

The intervals/blocks are sorted by their centers. The weights are exported in this order.

AdaptiveDensityApproximation.export_all — Function

export_all(grid)

Return one-dimensional arrays: centers, volumes, weights.

The arrays are sorted and exported by centers.

AdaptiveDensityApproximation.import_weights! — Function

import_weights!(grid, weights)

Import weights and return the mutated grid.

The intervals/blocks are sorted by their centers. The weights are imported in this order.

Refine a grid

AdaptiveDensityApproximation.refine! — Function

refine!(grid::Union{OneDimGrid,Grid}; block_variation::Function = default_block_variation, selection::Function = maximum, split_weights::Bool = false)

Subdivide intervals/blocks in a grid based on the respective variations. Return the mutated grid and the indices of subdivided (i.e. new) blocks (indices w.r.t arrays obtained from export_weights and export_all).

By default the variation of a block is the largest difference in weights w.r.t. to its neighbors. The blocks to be subdivided are those that have the largest variation. If several blocks have the same variation, all those blocks are subdivided.

Changing the selection of blocks

block_variation: Function to calculate the variation value for a block. Must use the following signature (block_center,block_volume, block_weight, neighbor_center,neighbor_volumes, neighbor_weights).

selection: Function to select the appropriate variation value(s) from. Must have the signature (variations) where variations is a one-dim array of the variation values.

Changing the weight splitting

By default, the subdividing blocks retain the weight of the original block. If split_weights = true, the weight of the original block is split up evenly between the subdividing blocks (i.e. divided by the number of subdividing blocks).

AdaptiveDensityApproximation.subdivide! — Function

subdivide!(grid::Union{OneDimGrid,Grid},block_name::AbstractString; split_weights::Bool = false)

Split the block with name block_name into 2^dim sub-blocks. Return the mutated grid.

Changing the weight splitting

By default, the subdividing blocks retain the weight of the original block. If split_weights = true, the weight of the original block is split up evenly between the subdividing blocks (i.e. divided by the number of subdividing blocks).

Restrict the grid domain

AdaptiveDensityApproximation.restrict_domain! — Function

restrict_domain!(grid::OneDimGrid; lower::Real = -Inf,upper::Real = Inf, weight_distribution::Symbol = :none)

Restrict a grid to the domain defined by lower and upper.

weight_distribution = :linear: A split block gets the weight scaled to the proportion of the block within the domain.

weight_distribution = :log: A split block gets the weight scaled to the proportion of the block, in a logarithmic scale, within the domain.

restrict_domain!(grid::Grid;lower = [-Inf,...,-Inf], upper = [Inf,...,Inf], weight_distribution::Symbol = :none)

Restrict a grid to the domain defined by lower and upper.

weight_distribution = :linear: A split block gets the weight scaled to the proportion of the block within the domain.

weight_distribution = :log: A split block gets the weight scaled to the proportion of the block, in a logarithmic scale, within the domain.

AdaptiveDensityApproximation.select_indices — Function

select_indices(grid::OneDimGrid; lower::Real=-Inf,upper::Real=Inf)

Return indices of intervals with centers between lower and upper. The indices correspond to the exported arrays centers, volumes and weights from export_weights and export_all.

select_indices(grid::Grid; lower= [-Inf,...,-Inf],upper=[Inf, ..., Inf])

Return indices of intervals with centers between lower and upper. The indices correspond to the exported arrays centers, volumes and weights from export_weights and export_all.

Simple calculations

Base.sum — Method

sum(grid::Union{OneDimGrid, Grid}; lower = nothing, upper = nothing, weight_distribution::Symbol = :none)

Return the sum of all weights.

lower, upper and weight_distribution can be used to restrict the domain, similar to restrict_domain!. This does not mutate the grid.

Too restrictive boundaries (empty grid) raise a warning and the default value 0 is returned.

Base.sum — Method

sum(f::Function,grid::Union{OneDimGrid, Grid}; lower = nothing, upper = nothing, weight_distribution::Symbol = :none)

Return the sum of f(weight) for all weights.

lower, upper and weight_distribution can be used to restrict the domain, similar to restrict_domain!. This does not mutate the grid.

Too restrictive boundaries (empty grid) raise a warning and the default value 0 is returned.

Base.prod — Method

prod(grid::Union{OneDimGrid, Grid}; lower = nothing, upper = nothing, weight_distribution::Symbol = :none)

Return the product of all weights.

lower, upper and weight_distribution can be used to restrict the domain, similar to restrict_domain!. This does not mutate the grid.

Too restrictive boundaries (empty grid) raise a warning and the default value 1 is returned.

Base.prod — Method

prod(f::Function,grid::Union{OneDimGrid, Grid}; lower = nothing, upper = nothing, weight_distribution::Symbol = :none)

Return the product of f(weight) for all weights.

lower, upper and weight_distribution can be used to restrict the domain, similar to restrict_domain!. This does not mutate the grid.

Too restrictive boundaries (empty grid) raise a warning and the default value 1 is returned.

AdaptiveDensityApproximation.integrate — Function

integrate(grid)

Return the sum of volume × weight for the intervals/blocks.

When the grid approximates a density φ with the weights of the intervals/blocks, integrate(grid) approximates the integral of the density over the grid domain: ∫_grid φ dV.

Integral models

AdaptiveDensityApproximation.integral_model — Function

integral_model(grid,f::Function, g::Function = f)

Create a model for the integral ∫_grid f(x,y,φ(y),...) dy. Returns

model function: (x,λ,args...) -> ∑_i block[i].volume × f(x,block[i].center,λ[i],args...)

initial parameter based on block weights: λ_0 = [block.weight for block in grid]

components of the sum as array of functions: [(x,λ,args...) -> block[i].volume × g(x,block[i].center,λ[i],args...) for i]

The functions f and g should have the arguments (x,center,weight,args...).

Partial derivatives

The optional function g can be used to obtain the partial derivatives of the model function w.r.t. λ as array of functions. For this, construct g such that

g(x,c,w) = ∂_w f(x,c,w)

Numeric PDF and CDF

AdaptiveDensityApproximation.get_pdf — Function

get_pdf(grid;normalize::Bool = true)

Return the discrete empirical pdf function of a grid. For this, the grid is understood as histogram, where the blocks are the bins, and the weights are the corresponding values. If normalize = true the values are normalized s.t. the sum of all values is 1.

AdaptiveDensityApproximation.get_cdf — Function

get_cdf(grid;normalize::Bool = true)

Return the discrete empirical cdf function of a grid. For this, the grid is understood as histogram, where the blocks are the bins, and the weights are the corresponding values. If normalize = true the values are normalized s.t. the sum of all values is 1.

Grid slices

AdaptiveDensityApproximation.get_slice — Function

get_slice(grid::Grid, slice_selection)

Return a grid of blocks (from grid) that intersect with a slice defined by slice_selection.

For example: the y-axis is defined by slice_selection = [0,nothing,0], the y-z-plane at x = 5 is defined by slice_selection = [5,nothing,nothing], etc..

AdaptiveDensityApproximation.dimension — Function

dimension(grid)

Return the dimension of the grid.